requirements.txt
pyTelegramBotAPI==4.21.0
vk_api==11.9.9
loguru==0.7.2
pytest==8.3.2
python-dotenv==1.0.1
